requirements/local.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/production.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r production.txt

# Deployment
Fabric

# Documentation
Sphinx

# Testing
nose
mock
django-nose
coverage
pep8
pylint

# Utility
django-mptt==0.6.0
django-extensions==1.3.3
django-taggit==0.11.2

# if you have issues on mavericks installing pycrypto use:
# ARCHFLAGS=-Wno-error=unused-command-line-argument-hard-error-in-future pip install pycrypto
